% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlB.R
\name{B.cor}
\alias{B.cor}
\title{Calculer une coefficient de correlation lineaire bayesien}
\usage{
B.cor(x, y, droite = 1, Niter = 55000, Nburn = 5000,
  message = TRUE)
}
\arguments{
\item{x, y}{Deux vecteurs de donnees quantitatives a correler}

\item{droite}{si egal a 1, la droite de correlation est ajoutee au graphique sinon seul le nuage de points est presente}

\item{Niter}{Le nombre d'iterations de la chaine MCMC}

\item{Nburn}{Le nombre d'iterations a bruler de la chaine MCMC}

\item{message}{}
}
\description{
Calculer une coefficient de correlation lineaire bayesien
}
\examples{
x <- rnorm(30, 20, 5)
y <- 2 * x + rnorm(30, 0,10)
B.cor(x,y)
}
