% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.mean.R
\name{B.mean}
\alias{B.mean}
\title{Comparaison bayesienne de deux moyennes Gaussiennes}
\usage{
B.mean(X1, X2, n.iter = 11000, n.burnin = 1000, n.thin = 1,
  prior.norm = c(0, 0.001, 0, 0.001), prior.gamma = c(0.5, 0.5, 0.5,
  0.5))
}
\arguments{
\item{X1, X2}{Deux vecteurs Gaussiens}

\item{n.iter}{Le nombre d'iterations de la cha?ne MCMC}

\item{n.burnin}{Le nombre d'iterations a bruler de la chaine MCMC}

\item{n.thin}{Le nombre d'iterations a filtrer de la chaine MCMC}

\item{prior.norm}{Un vecteur de deux valeurs a priori sur les moyennes}

\item{prior.gamma}{Un vecteur de deux valeurs a priori sur les parametres de precision (Distribution Gamma)}
}
\description{
Comparaison bayesienne de deux moyennes par methodes MCMC
}
\examples{
X1 <- rnorm(15,6,1)
X2 <- rnorm(10,5,.5)
B.mean(X1,X2)
}
