% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.moy1.R
\name{B.moy1}
\alias{B.moy1}
\title{Utile pour B.mean}
\usage{
B.moy1(var1, n.iter = 11000, n.burnin = 1000, n.thin = 1,
  prior.norm = c(0, 0.001, 0, 0.001), prior.gamma = c(0.5, 0.5, 0.5,
  0.5))
}
\arguments{
\item{A}{definir}
}
\description{
Utile pour B.mean
}
