% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.moy2.R
\name{B.moy2}
\alias{B.moy2}
\title{Utile pour B.mean}
\usage{
B.moy2(var1, var2, n.iter = 11000, n.burnin = 1000, n.thin = 1,
  prior.norm = c(0, 0.001, 0, 0.001), prior.gamma = c(0.5, 0.5, 0.5,
  0.5))
}
\arguments{
\item{A}{definir}
}
\description{
Utile pour B.mean
}
\examples{
aa <- rnorm(30,mean = 100,sd = 20)
bb <- rnorm(30,mean = 90,sd = 20)
t.test(aa,bb)
res <- B.moy2(aa,bb,n.iter=101000)
mean(res$sims.list$p)
hist(res$sims.list$di)
}
