% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.pois.R
\name{B.pois}
\alias{B.pois}
\title{Comparaison Bayesienne de deux moyennes de comptage}
\usage{
B.pois(X1, X2, n.iter = 11000, n.burnin = 1000, n.thin = 1,
  prior.gamma = c(0.1, 0.1, 0.1, 0.1))
}
\arguments{
\item{X1, X2}{Deux vecteurs de donnees de Comptage}

\item{n.iter}{Le nombre d'iterations de la chaine MCMC}

\item{n.burnin}{Le nombre d'iterations a bruler de la cha?ne MCMC}

\item{n.thin}{Le nombre d'iterations a filtrer de la chaine MCMC}

\item{prior.gamma}{Un vecteur de deux valeurs ? priori sur les parametres de la distribution Gamma pour les Lambda (Disutribution Poisson)}
}
\description{
Comparaison Bayesienne de deux moyennes de comptage (issus de distributions de Poisson)
}
\examples{
X1 <- rpois(n = 25, lambda = 2)
X2 <- rpois(n = 25, lambda = 2)
B.pois(X1,X2)
}
