% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.pois1.R
\name{B.pois1}
\alias{B.pois1}
\title{Utile pour B.pois}
\usage{
B.pois1(var1, n.iter = 11000, n.burnin = 1000, n.thin = 1,
  prior.gamma = c(0.1, 0.1, 0.1, 0.1))
}
\arguments{
\item{A}{definir}
}
\description{
Utile pour B.pois
}
