% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.pois2.R
\name{B.pois2}
\alias{B.pois2}
\title{Utile pour B.pois}
\usage{
B.pois2(var1, var2, n.iter = 11000, n.burnin = 1000, n.thin = 1,
  prior.gamma = c(0.1, 0.1, 0.1, 0.1))
}
\arguments{
\item{A}{definir}
}
\description{
Utile pour B.pois
}
