% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.prop.R
\name{B.prop}
\alias{B.prop}
\title{Comparaison Bayesienne de deux proportions par methodes MCMC}
\usage{
B.prop(X1, X2, n.iter = 11000, n.burnin = 1000, n.thin = 1,
  prior.beta = c(1, 1, 1, 1))
}
\arguments{
\item{X1, X2}{Deux vecteurs de donnees binaires codees 0/1}

\item{n.iter}{Le nombre d'iterations de la chaine MCMC}

\item{n.burnin}{Le nombre d'iterations a bruler de la cha?ne MCMC}

\item{n.thin}{Le nombre d'iterations a filtrer de la chaine MCMC}

\item{prior.beta}{Un vecteur (a,b,c,d) de deux valeurs a priori sur les distributions Beta (a,b) et (c,d)}
}
\description{
Comparaison Bayesienne de deux proportions
}
\examples{
X1 <- rbinom(n = 10,size = 1,prob = .4)
X2 <- rbinom(n = 12,size = 1,prob = .6)
B.prop(X1,X2)
}
