% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.prop1.R
\name{B.prop1}
\alias{B.prop1}
\title{Utile pour B.prop}
\usage{
B.prop1(var1, n.iter = 11000, n.burnin = 1000, n.thin = 1,
  prior.beta = c(1, 1, 1, 1))
}
\arguments{
\item{A}{definir}
}
\description{
Utile pour B.prop
}
