% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burt.R
\name{burt}
\alias{burt}
\title{Make a disjonctif table}
\usage{
burt(x)
}
\arguments{
\item{x}{a factor vector}
}
\description{
Make a disjonctif table#'
}
\details{
Does the same as tab.disjonctif function from the  FactoMineR package (but slower :-) and only on vectors)
}
\examples{
fact <-sample(x = LETTERS[1:5], size=10, replace=T)
burt(fact)
require(FactoMineR)
tab.disjonctif(dtf)
}
\author{
JuG
}
