% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correl.r
\name{correl}
\alias{correl}
\title{Do something}
\usage{
correl(x, y, droite = 1, nomx = NULL, nomy = NULL)
}
\description{
Do something
}
\examples{
x <- rnorm(n=20, mean=10)
y <- rnorm(n=20, mean=15)
z <- 2 * x + rnorm(n=20, sd=2)
correl(x,y)
correl(x,z)
}
\author{
JuG
}
