% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.r
\name{cs}
\alias{cs}
\title{Scale and reduce data}
\usage{
cs(x)
}
\arguments{
\item{x}{a numeric vector to be scaled and reduced}
}
\description{
Scale and reduce data. Does exactly the same as scale but slightly faster.
}
\examples{
x <-rnorm(n = 100, mean = 50, sd = 10)
xreduced <- cs(x)
c(mean(xreduced), sd(xreduced))
xscaled <- scale(x)
c(mean(xscaled), sd(xscaled))

x <-rnorm(n = 1e8, mean = 50, sd = 10)
system.time(c(mean(cs(x)), sd(cs(x))))
system.time(c(mean(scale(x)), sd(scale(x))))
}
