% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc1.r
\name{desc1}
\alias{desc1}
\title{Descriptive statistics of a given variable}
\usage{
desc1(x, modmax = 7)
}
\arguments{
\item{x}{}
}
\description{
Descriptive statistics of a given variable
}
\examples{
x <- runif(100)
desc1(x)
y <- gl(n=2, k=30,labels = letters[1:2])
desc1(y)
}
\author{
JuG
}
