% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descd.r
\name{descd}
\alias{descd}
\title{Do something}
\usage{
descd(D)
}
\description{
TODO voir ce que définit Var textuelles
}
\examples{
dtf <- data.frame(a =rnorm(20), 
                  b= sample(x = LETTERS[1:5], size  = 20,replace=T), 
                  c = sample(x = LETTERS[1:5], size  = 20,replace=T),
                  d = rep(date(),20))
dtf$c <- as.character(dtf$c)
descd(dtf)

}
\author{
JuG
}
