% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr1.r
\name{descr1}
\alias{descr1}
\title{Descriptive statistics for quantitative data}
\usage{
descr1(Y, Tap = FALSE)
}
\arguments{
\item{Y}{numeric vector}

\item{Tap}{boolean (tri à plat)}
}
\description{
Descriptive statistics for quantitative data
}
\examples{
Y <- runif(n = 100)
descr1(Y)
}
\author{
JuG
}
