% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr3.r
\name{descr3}
\alias{descr3}
\title{Descriptive statistics for quantitative data as a function of a qualitative data}
\usage{
descr3(Y, X, Tap = FALSE, nom = NULL, nomY = NULL, latex = 0)
}
\arguments{
\item{Y}{numeric vector}

\item{X}{factor (or group) vector}

\item{Tap}{boolean (tri à plat)}

\item{nom}{X label}

\item{nomY}{Y label}

\item{latex}{boolean latex output}
}
\description{
Descriptive statistics for quantitative data as a function of a qualitative data
}
\examples{
X <- sample(x = letters[1:2],size = 100, replace = T)
Y <- runif(n = 100)
descr3(Y, X)
descr3(Y, X, latex = T)
}
\author{
JuG
}
