% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme.r
\name{extreme}
\alias{extreme}
\title{Do something}
\usage{
extreme(x, nlab = 2, labs = as.character(1:length(x)),
  ylab = "Sorted Data", ...)
}
\description{
NE fonctionne pas avec valeurs extemement faibles?
}
\examples{
x <- rnorm(100, mean = 20, sd=2)
x[10] <- 100
extreme(x)
x[3] <- 3
extreme(x)
}
\author{
JuG
}
