% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggboxplot.r
\name{ggboxplot}
\alias{ggboxplot}
\title{Do something}
\usage{
ggboxplot(x, y, Groupe = NULL)
}
\description{
Do something
}
\examples{
x <- rnorm(n=20, mean=10)
y <- rep(LETTERS[1:2], each=20)
grp <- rep(LETTERS[3:4], 20)
ggboxplot( x=x, y=y)
ggboxplot( x=x, y=y, Groupe = grp)
}
\author{
JuG
}
