% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcompar.r
\name{ggcompar}
\alias{ggcompar}
\title{Do something}
\usage{
ggcompar(x, y, DDD = NULL)
}
\arguments{
\item{x}{a string corresponding to the name of the column in the data frame}

\item{y}{a string corresponding to the name of the column in the data frame}

\item{DDD}{a data frame}
}
\description{
Do something
}
\examples{
x <- c(rnorm(n=30, mean=5), rnorm(n=30, mean=10))
grp <- rep(LETTERS[1:2], each=30)
dtf <- data.frame(x, grp)
ggcompar(x = "x", y = "grp", DDD = dtf)

}
\author{
JuG
}
