% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghist.r
\name{gghist}
\alias{gghist}
\title{Do something}
\usage{
gghist(x1, Groupe = NULL)
}
\description{
Do something
}
\examples{
x <- c(rnorm(n=10, mean=5), rnorm(n=10, mean=10))
grp <- rep(LETTERS[1:2], each=10)
gghist(x1 = x)
gghist(x1 = x, Groupe = grp)
}
\author{
JuG
}
