% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpoints.r
\name{ggpoints}
\alias{ggpoints}
\title{Do something}
\usage{
ggpoints(x, y, droite = FALSE, nomx = NULL, nomy = NULL)
}
\description{
Do something
}
\examples{
x <- rnorm(n=20, mean=10)
y <- rnorm(n=20, mean=15)
ggpoints(x,y, droite = T, nomx = "X-axis")

}
\author{
JuG
}
