% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurv.r
\name{ggsurv}
\alias{ggsurv}
\title{Do something}
\usage{
ggsurv(s, CI = "def", plot.cens = T, surv.col = "gg.def",
  cens.col = "red", lty.est = 1, lty.ci = 2, cens.shape = 3,
  back.white = F, xlab = "Time", ylab = "Survival", main = "")
}
\description{
Do something
}
\examples{


}
\author{
JuG
}
