% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurvie.r
\name{ggsurvie}
\alias{ggsurvie}
\title{Do something}
\usage{
ggsurvie(x, y, groups = 0, latex = 0, titre = 0)
}
\description{
Do something
}
\examples{
status <- rbinom(n=50, size=1, prob = .5)
time <- rexp(n=50) * 20
ggsurvie(x=time, y=status)
}
\author{
JuG
}
