% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmB.R
\name{glmB}
\alias{glmB}
\title{Ajuster un modele de type GLM ou GLMER par MCMC}
\usage{
glmB(formula, data = NULL, family = "gaussian", Niter = 105000,
  Nburnin = 5000, Nthin = 5, Prior = NULL, Seuilmonit = NULL,
  Gammaalea = c(0.1, 0.1), Diagno = TRUE, logiciel = "jags",
  dig = 3, Graph = TRUE, Alea = FALSE)
}
\arguments{
\item{formula}{une formule de predicteur lineaire, eventuellement mixte (+1|sujet)}

\item{data}{la base de donnees a utiliser}

\item{family}{la famille du GLM (gaussian,binomial,poisson,gamma et beta(NON GLM))}

\item{Niter}{Le nombre d'iterations souhaite pour la chaine MCMC}

\item{Nburnin}{Le nombre d'iterations a bruler pour la chaine MCMC}

\item{Nthin}{Le nombre d'iterations a filtrer pour la chaine MCMC}

\item{Prior}{Une matice de k colonnes et 2 lignes, de moyennes et precisions (1/var) a priori (pour chaque coef)}

\item{Seuilmonit}{Un vecteur de x elements, (x parametres), si on souhaite monitorer la Proba d'?tre sup au seuil}

\item{Gammaalea}{Un vecteur de 2 elements, moyennes et variances a priori de l'effet aleatoire}

\item{dig}{Nombre de decimales pour la presentation des resultats}

\item{Graph}{TRUE par defaut, graphiques des distributions a posteriori}

\item{Alea}{FALSE par defaut. Si TRUE: monitore et affiche les effets aleatoires}

\item{diagno}{Uniquement pour le modele logistique(et mixte), si on souhaite le detail des diagnostiques}
}
\description{
Fonction permettant d'ajuster un modele lineaire generalise en bayesien a l'aide de methodes MCMC
}
\examples{
dtf <- data.frame(Y = rbinom(n = 60 ,size=1,prob = .3), X = rnorm(60, 10,2) )
glmB(Y ~ X, data= dtf, family = "binomial")

}
