% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmB_gamma_JAGS.R
\name{glmB_gamma_JAGS}
\alias{glmB_gamma_JAGS}
\title{Utile pour glmB
 @description  Utile pour glmB}
\usage{
glmB_gamma_JAGS(formul, data = NULL, Niter = 105000, Nburnin = 5000,
  Nthin = 5, Prior = NULL, Seuilmonit = NULL, Graph = TRUE)
}
\description{
Utile pour glmB
 @description  Utile pour glmB
}
