% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmB_logit_JAGS.R
\name{glmB_logit_JAGS}
\alias{glmB_logit_JAGS}
\title{Utile pour glmB}
\usage{
glmB_logit_JAGS(formul, data = NULL, Niter = 105000, Nburnin = 5000,
  Nthin = 5, Prior = NULL, Seuilmonit = NULL, Diagno = TRUE,
  Graph = TRUE)
}
\description{
Utile pour glmB
}
