% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmB_poisson_JAGS.R
\name{glmB_poisson_JAGS}
\alias{glmB_poisson_JAGS}
\title{Utile pour glmB}
\usage{
glmB_poisson_JAGS(formul, data = NULL, Niter = 105000,
  Nburnin = 5000, Nthin = 5, Prior = NULL, Seuilmonit = NULL,
  Graph = TRUE)
}
\description{
Utile pour glmB
}
