\name{gmrcfun-package}
\alias{gmrcfun-package}
\alias{gmrcfun}
\docType{package}
\title{
Fonctions GMRC HUS
}
\description{
Ensemble de fonctions utiles au GMRC HUS Strasbourg pour le traitement statistique de donnees medicales
}
\details{
\tabular{ll}{
Package: \tab gmrcfun\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-07\cr
License: \tab IV\cr
}
Possibilites d'utiliser les fonctions glmB, normal, descr1, tableaux, etc...
}
\author{
Mickael Schaeffer

Maintainer: Mickael Schaeffer <mickael.schaeffer@chru-strasbourg.fr>
}
\references{
Aucune reference}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ jags ~~
}
\examples{
~~ glmB(Y~X,family="binomial"), normal(c(1.5,1.54,1.48,1.50,1.51)) ~~
}
