% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HosLem.r
\name{HosLem}
\alias{HosLem}
\title{The Hosmer-Lemeshow goodness of fit test for logistic regression}
\usage{
HosLem(modele)
}
\description{
Hosmer-Lemeshow goodness of fit test for logistic regression.
}
\details{
If the p-value is small, this is indicative of poor fit
}
\examples{


}
\author{
JuG
}
