% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.diff.prop.R
\name{IC.diff.prop}
\alias{IC.diff.prop}
\title{Intervalle de confiance d'une difference de deux proportions}
\usage{
IC.diff.prop(x1, n1, x2, n2, alpha01 = 0.5, alpha02 = 0.5,
  beta01 = 0.5, beta02 = 0.5, val = 0.95)
}
\arguments{
\item{x1}{Succes groupe 1}

\item{n1}{Patients groupe 1}

\item{x2}{Succes groupe 2}

\item{n2}{Patients groupe 2}

\item{alpha01}{parametre a priori sur la distribution beta de la premiere proportion}

\item{alpha02}{parametre a priori sur la distribution beta de la deuxieme proportion}

\item{beta01}{parametre a priori sur la distribution beta de la premiere proportion}

\item{beta02}{parametre a priori sur la distribution beta de la deuxieme proportion}

\item{val}{largeur de l'intervalle de credibilite, par defaut 0.95}
}
\value{
Bayesian, Wald, continuity correction Wald and approx Bayesian Confidence Intervals
}
\description{
Intervalle de confiance d'une difference de deux proportions
}
\examples{
IC.diff.prop(x1=2,n1=10,x2=3,n2=11)
}
