% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names2.r
\name{names2}
\alias{names2}
\title{lister les variables d'un jeu de donnees avec les numeros a cote}
\usage{
names2(D)
}
\description{
lister les variables d'un jeu de donnees avec les numeros a cote
}
\examples{


}
\author{
JuG
}
