% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.evol.r
\name{plot.evol}
\alias{plot.evol}
\title{Do something}
\usage{
\method{plot}{evol}(Trajectoires, Groupe = NULL, IC = FALSE,
  Moyenne = TRUE, Temps = NULL, label_x = "Temps",
  label_y = "Valeur", labels_ticks_x = NULL)
}
\description{
Do something
}
\examples{


}
\author{
JuG
}
