% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roc.r
\name{Roc}
\alias{Roc}
\title{ROC curve for logistic regression}
\usage{
Roc(mod)
}
\arguments{
\item{mod}{logistic regression model}
}
\description{
ROC curve for logistic regression
}
\examples{
dtf <- data.frame(Y = rbinom(n = 60 ,size=1,prob = .3), X = rnorm(60, 10,2) )
mod <- glm(Y ~ X, data= dtf, family = "binomial")
Roc(mod)
}
\author{
JuG
}
