% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableaux.r
\name{tableaux}
\alias{tableaux}
\title{Do something}
\usage{
tableaux(X, Y, latex = 0, p.val = T, entete = 0, caption = 0)
}
\arguments{
\item{Y}{}
}
\description{
Do something
}
\examples{
X <- sample(x = letters[1:2],size = 20, replace = T)
Y <- sample(x = letters[3:4],size = 20, replace = T)
tableaux(X,Y, latex=T)
}
\author{
JuG
}
